refactor: improve delegate auth flow and token handling
- Change RequestId type from uint64_t to uint32_t for consistency
- Remove ActivateToken and RevokeTokens methods from CompanionManagerImpl
- Update FFI interface: restructure data arrays and remove unused commands
- Replace GetActiveUserIdManager with GetUserIdManager
- Fix log format specifiers (PRIu64 to u) for RequestId changes
- Replace magic numbers with named constants in test code
- Update unit tests to reflect API changes

+++ services/common/inc/service_common.h
@@ -35,7 +35,7 @@ namespace CompanionDeviceAuth {
 
 using FwkResultCallback = std::function<void(ResultCode result, const std::vector<uint8_t> &extraInfo)>;
 
-using RequestId = uint64_t;
+using RequestId = uint32_t;
 using TemplateId = uint64_t;
 using ScheduleId = uint64_t;
 using UserId = int32_t;
